Support IPv6-only network segments
The primary change in this commit is to add a subnet that uses IPv6
exclusively.

Frustratingly, several external services (Github, for
example) do not support IPv6 yet. In order to reach IPv4-only services
from an IPv6-only resource, we need to enable (A) DNS64 - a service that
resolves IPv4 to an AWS issued IPv6 with a specific pattern, and
(B) NAT64 - a way to reverse translate DNS64 issued IPv6 to the original
IPv4.

This also requires the use of a NAT gateway and an elastic IP for the NAT
gateway - which also subjects the account holder to additional costs
associated with traffic that is passed through the NAT gateway. This
sucks, but it is the state of networking at AWS at the time of this
commit.

- Add IPv6-only subnets and associate private route table to them
- Add NAT gateway to route DNS64-issued addresses through
- Add Elastic IP for NAT gateway
- Add Route Table entries for DNS64-issued addresses
- Add outputs in example module

If you are reading this and are employed at Github or Gitlab, please
lobby internally for IPv6 support for your services!